Large refactoring of generation of full word shape (part 1)
This commit is the first part of a large refactoring of the way
the screen is managed.

Until now, the screen was divided into 5x3 cells, one per letter.
This does not work well with letter with various size, and does not
allow global word deformations.

The new approach does not have cell: the screen is a continuous space.

The preparation of the path describing a word is first managed by a
TextShaper that scale the letters, and ensure they are visually
connected inside a word, and then by a ScreenManager which places the
whole word on the screen (by simply setting the ShapedWord.origin offset).

Current state:
--------------

The TextShaper is usable. It takes care of the generation of the final
letters' shapes, thus removing some code from learning_words_nao.py.

The ScreenManager is only a stub which currently always places
words in the center of the screen. The placing need to be more
intelligent, and more importantly, *it needs to be able to recognize for
which letter a demonstration is given*: until then, the CoWriter app is
mostly useless since one can not provide feedback.